API

WrappedMultiGraphs.asmultiedges — Method

asmultiedges(edgs) -> Any

Get a collection of SingleMultiEdge from the edgs

WrappedMultiGraphs.getgraph — Method

getgraph(mg::WrappedMultiGraphs.AbstractMultiGraph) -> Any

Get the wrapped graph

WrappedMultiGraphs.multiplicity — Method

multiplicity(
    mg::MultiDiGraph,
    e::Graphs.SimpleGraphs.SimpleEdge
) -> Int64

Get the multiplicity, i.e. the number, of the edge e

WrappedMultiGraphs.multiplicity — Method

multiplicity(
    mg::MultiGraph,
    s::Integer,
    d::Integer
) -> Int64

Get the multiplicity, i.e. the number, of an edge with source s and destination d.

WrappedMultiGraphs.todirected — Method

todirected(mg::MultiGraph) -> MultiDiGraph

Convert to a directed graph

WrappedMultiGraphs.toundirected — Method

toundirected(mg::MultiDiGraph) -> MultiGraph

Convert to an undirected graph

WrappedMultiGraphs.MultiDiGraph — Type

mutable struct MultiDiGraph{T<:Integer} <: WrappedMultiGraphs.AbstractMultiGraph{T<:Integer}

graph::SimpleDiGraph: The wrapped graph

WrappedMultiGraphs.MultiGraph — Type

struct MultiGraph{T<:Integer} <: WrappedMultiGraphs.AbstractMultiGraph{T<:Integer}

graph::SimpleGraph: The wrapped graph

WrappedMultiGraphs.SingleMultiEdge — Type

struct SingleMultiEdge{T<:Integer} <: Graphs.SimpleGraphs.AbstractSimpleEdge{T<:Integer}

src::Integer: The source node

dst::Integer: The destination node

m::Integer: The multiplicity of the edge

The m-th instance of an edge connecting src and dst.

Index

WrappedMultiGraphs.MultiDiGraph

WrappedMultiGraphs.MultiGraph

WrappedMultiGraphs.SingleMultiEdge

WrappedMultiGraphs.asmultiedges

WrappedMultiGraphs.getgraph

WrappedMultiGraphs.multiplicity

WrappedMultiGraphs.multiplicity

WrappedMultiGraphs.todirected

WrappedMultiGraphs.toundirected