API

WMultiGraphs.asmultiedges — Method

asmultiedges(edgs) -> Any

Get a collection of SingleMultiEdge from the edgs

WMultiGraphs.getgraph — Method

getgraph(mg::WMultiGraphs.AbstractMultiGraph) -> Any

Get the wrapped graph

WMultiGraphs.multiplicity — Method

multiplicity(
    mg::MultiDiGraph,
    e::Graphs.SimpleGraphs.SimpleEdge
) -> Int64

Get the multiplicity, i.e. the number, of the edge e

WMultiGraphs.multiplicity — Method

multiplicity(
    mg::MultiGraph,
    s::Integer,
    d::Integer
) -> Int64

Get the multiplicity, i.e. the number, of an edge with source s and destination d.

WMultiGraphs.todirected — Method

todirected(mg::MultiGraph) -> MultiDiGraph

Convert to a directed graph

WMultiGraphs.toundirected — Method

toundirected(mg::MultiDiGraph) -> MultiGraph

Convert to an undirected graph

WMultiGraphs.MultiDiGraph — Type

mutable struct MultiDiGraph{T<:Integer} <: WMultiGraphs.AbstractMultiGraph{T<:Integer}

graph::SimpleDiGraph: The wrapped graph

WMultiGraphs.MultiGraph — Type

struct MultiGraph{T<:Integer} <: WMultiGraphs.AbstractMultiGraph{T<:Integer}

graph::SimpleGraph: The wrapped graph

WMultiGraphs.SingleMultiEdge — Type

struct SingleMultiEdge{T<:Integer} <: Graphs.SimpleGraphs.AbstractSimpleEdge{T<:Integer}

src::Integer: The source node

dst::Integer: The destination node

m::Integer: The multiplicity of the edge

The m-th instance of an edge connecting src and dst.

Index

WMultiGraphs.MultiDiGraph

WMultiGraphs.MultiGraph

WMultiGraphs.SingleMultiEdge

WMultiGraphs.asmultiedges

WMultiGraphs.getgraph

WMultiGraphs.multiplicity

WMultiGraphs.multiplicity

WMultiGraphs.todirected

WMultiGraphs.toundirected